feat: refactor existing redirects
