+++ blog-website/api/fallback/redirectsBacklinks.js
@@ -1,0 +1,286 @@
+const brokenBacklinkGeneratedRedirects = [
+  {
+    route: '/2019/10/definitely-typed-movie.html',
+    redirect: '/definitely-typed-the-movie',
+  },
+  {
+    route: '/2019/10/08/definitely-typed-movie',
+    redirect: '/definitely-typed-the-movie',
+  },
+  {
+    route: '/2019/06/typescript-webpack-you-down-with-pnp.html',
+    redirect: '/2019/06/07/typescript-webpack-you-down-with-pnp',
+  },
+  {
+    route:
+      '/2019/07/typescript-and-eslint-meet-fork-ts-checker-webpack-plugin.html',
+    redirect:
+      '/2019/07/13/typescript-and-eslint-meet-fork-ts-checker-webpack-plugin',
+  },
+  {
+    route: '/2012/04/making-pdfs-from-html-in-c-using.html',
+    redirect: '/2012/04/05/making-pdfs-from-html-in-c-using',
+  },
+  {
+    route: '/2012/09/globalize-and-jquery-validate.html/',
+    redirect: '/2012/09/06/globalize-and-jquery-validate',
+  },
+  {
+    route: '/2014/05/typescript-jsdoc-and-intellisense.html/',
+    redirect: '/2014/05/05/typescript-jsdoc-and-intellisense',
+  },
+  {
+    route: '/2015/02/using-gulp-in-asp-net-instead-of-web-optimization.html/',
+    redirect: '/2015/02/17/using-gulp-in-asp-net-instead-of-web-optimization',
+  },
+  {
+    route: '/2015/10/jquery-validation-globalize-hits-10.html/',
+    redirect: '/2015/10/05/jquery-validation-globalize-hits-10',
+  },
+  {
+    route: '/2015/12/es6-typescript-babel-react-flux-karma.html/',
+    redirect: '/2015/12/16/es6-typescript-babel-react-flux-karma',
+  },
+  {
+    route: '/2016/05/the-mysterious-case-of-webpack-angular-and-jquery.html/',
+    redirect: '/2016/05/24/the-mysterious-case-of-webpack-angular-and-jquery',
+  },
+  {
+    route: '/2016/11/my-subconscious-is-better-developer.html/',
+    redirect: '/2016/11/12/my-subconscious-is-better-developer',
+  },
+  {
+    route: '/2017/02/typescript-types-and-repeatable-builds.html/',
+    redirect: '/2017/02/14/typescript-types-and-repeatable-builds',
+  },
+  {
+    route: '/2017/06/windows-defender-step-away-from-npm.html',
+    redirect: '/2017/06/11/windows-defender-step-away-from-npm',
+  },
+  {
+    route: '/2017/11/the-typescript-webpack-pwa.html/',
+    redirect: '/2017/11/19/the-typescript-webpack-pwa',
+  },
+  {
+    route: '/2018/01/finding-webpack-4-use-map.html/',
+    redirect: '/2018/01/29/finding-webpack-4-use-map',
+  },
+  {
+    route: '/2018/03/its-not-dead-webpack-and-dead-code.html/',
+    redirect: '/2018/03/07/its-not-dead-webpack-and-dead-code',
+  },
+  {
+    route: '/2018/07/cypress-and-auth0.html/',
+    redirect: '/2018/07/09/cypress-and-auth0',
+  },
+  {
+    route: '/2018/12/you-might-not-need-thread-loader.html/',
+    redirect: '/2018/12/22/you-might-not-need-thread-loader',
+  },
+  {
+    route: '/2019/04/react-select-with-less-typing-lag.html/',
+    redirect: '/2019/04/27/react-select-with-less-typing-lag',
+  },
+  {
+    route: '/2019/05/typescript-and-high-cpu-usage-watch.html/',
+    redirect: '/2019/05/23/typescript-and-high-cpu-usage-watch',
+  },
+  {
+    route:
+      '/2020/01/ef-core-31-breaks-left-join-with-no-navigation-property.html/',
+    redirect:
+      '/2020/01/02/ef-core-31-breaks-left-join-with-no-navigation-property',
+  },
+  {
+    route: '/2020/01/license-to-kill-your-pwa.html/',
+    redirect: '/2020/01/21/license-to-kill-your-pwa',
+  },
+  {
+    route: '/2020/01/from-create-react-app-to-pwa.html/',
+    redirect: '/2020/01/31/from-create-react-app-to-pwa',
+  },
+  {
+    route: '/2020/02/web-workers-comlink-typescript-and-react.html/',
+    redirect: '/2020/02/21/web-workers-comlink-typescript-and-react',
+  },
+  {
+    route: '/2020/03/dual-boot-authentication-with-aspnetcore.html/',
+    redirect: '/2020/03/22/dual-boot-authentication-with-aspnetcore',
+  },
+  {
+    route: '/2020/03/offline-storage-in-pwa.html/',
+    redirect: '/2020/03/29/offline-storage-in-pwa',
+  },
+  {
+    route: '/2020/05/from-react-window-to-react-virtual.html/',
+    redirect: '/2020/05/10/from-react-window-to-react-virtual',
+  },
+  {
+    route: '/2020/06/taskwhenall-select-is-footgun.html/',
+    redirect: '/2020/06/21/taskwhenall-select-is-footgun',
+  },
+  {
+    route: '/2020/10/autofac-6-integration-tests-and-generic-hosting.html/',
+    redirect: '/2020/10/02/autofac-6-integration-tests-and-generic-hosting',
+  },
+  {
+    route: '/2020/11/bulletproof-uniq-with-typescript.html/',
+    redirect: '/2020/11/14/bulletproof-uniq-with-typescript',
+  },
+  {
+    route: '/2020/12/nullable-reference-types-csharp-strictnullchecks.html/',
+    redirect: '/2020/12/20/nullable-reference-types-csharp-strictnullchecks',
+  },
+  {
+    route:
+      '/2020/12/prettier-your-csharp-with-dotnet-format-and-lint-staged.html/',
+    redirect:
+      '/2020/12/22/prettier-your-csharp-with-dotnet-format-and-lint-staged',
+  },
+  {
+    route: '/2021/01/create-react-app-with-ts-loader-and-craco.html/',
+    redirect: '/2021/01/02/create-react-app-with-ts-loader-and-craco',
+  },
+  {
+    route: '/2021/01/strongly-typing-react-query-s-usequeries.html/',
+    redirect: '/2021/01/03/strongly-typing-react-query-s-usequeries',
+  },
+  {
+    route: '/2021/01/azure-easy-auth-and-roles-with-dotnet-and-core.html/',
+    redirect: '/2021/01/14/azure-easy-auth-and-roles-with-dotnet-and-core',
+  },
+  {
+    route: '/2021/01/aspnet-serilog-and-application-insights.html/',
+    redirect: '/2021/01/30/aspnet-serilog-and-application-insights',
+  },
+  {
+    route:
+      '/2021/02/goodbye-client-affinity-hello-data-protection-with-azure.html/',
+    redirect:
+      '/2021/02/27/goodbye-client-affinity-hello-data-protection-with-azure',
+  },
+  {
+    route: '/2020/10/31/azure-devops-node-api-missing-episodes/',
+    redirect: '/2020/10/31/azure-devops-node-api-git-api-getrefs-wiki-api',
+  },
+  {
+    route: '/2020/10/azure-devops-node-api-missing-episodes.html/',
+    redirect: '/2020/10/31/azure-devops-node-api-git-api-getrefs-wiki-api',
+  },
+  {
+    route:
+      '/assets/images/i-must-break-you-cc86f4365eea8125204d16f065712ec9.jpg',
+    redirect:
+      '/assets/images/i-must-break-you-6212db0b9e54e386eb08a58d0e170dd1.webp',
+  },
+  {
+    route: '/2020/05/21/autofac-webapplicationfactory-and/',
+    redirect: '2020/05/21/autofac-webapplicationfactory-integration-tests',
+  },
+  {
+    route: '/img/favicon.ico',
+    redirect: '/favicon.ico',
+  },
+  {
+    route: '/apple-touch-icon-precomposed.png',
+    redirect: '/apple-touch-icon.png',
+  },
+  {
+    route: '/img/favicon/apple-touch-icon.png',
+    redirect: '/apple-touch-icon.png',
+  },
+  {
+    route: '/2019/08/hard-coding-claim-in-development-mode.html',
+    redirect: '/2019/08/02/asp-net-authentication-hard-coding-claims',
+  },
+  {
+    route: '/2021/03/15/from-blogger-to-docusaurus/',
+    redirect: '/definitive-guide-to-migrating-from-blogger-to-docusaurus',
+  },
+  {
+    route: '/2021/01/surfacing-azure-pipelines-build-info-in.html',
+    redirect:
+      '/2021/01/29/surfacing-azure-pipelines-build-info-in-an-aspnet-react-app',
+  },
+  {
+    route: '/2021/02/generate-typescript-and-csharp-clients-with-nswag.html',
+    redirect: '/2021/03/06/generate-typescript-and-csharp-clients-with-nswag',
+  },
+  {
+    route: '/2014/12/blog.icanmakethiswork.io',
+    redirect: '/2014/12/05/whats-in-a-name',
+  },
+  {
+    route: '/sitemap',
+    redirect: '/sitemap.xml',
+  },
+  {
+    route: '/sitemap.txt',
+    redirect: '/sitemap.xml',
+  },
+  {
+    route: '/2019/03/06/the-big-one-point-oh',
+    redirect: '/2019/03/06/fork-ts-checker-webpack-plugin-v1',
+  },
+  {
+    route: '/2019/03/06/the-big-one-point-oh.html',
+    redirect: '/2019/03/06/fork-ts-checker-webpack-plugin-v1',
+  },
+  {
+    route: '/2021/03/managed-identity-azure-sql-and-entity.html',
+    redirect: '/2021/03/10/managed-identity-azure-sql-entity-framework',
+  },
+  {
+    route: '/2018/08/killing-relative-paths-with-typescript-and.html',
+    redirect: '/2018/08/21/typescript-webpack-alias-goodbye-relative-paths',
+  },
+  {
+    route: '/2020/05/21/autofac-webapplicationfactory-and',
+    redirect: '/2020/05/21/autofac-webapplicationfactory-integration-tests',
+  },
+  {
+    route: '/tags/azure-devops-api',
+    redirect: '/tags/azure-dev-ops-api',
+  },
+  {
+    route: '/2019/12/18/automating-teams-notifications-recently',
+    redirect: '/2019/12/18/teams-notification-webhooks',
+  },
+  {
+    route: '/2018/07/configuring-docker-azure-web-app-containers.html',
+    redirect:
+      '/2018/07/28/azure-app-service-web-app-containers-asp-net-nested-configuration',
+  },
+  {
+    route: '/2019/02/whitelist-proxying-with-aspnet-core.html',
+    redirect: '/2019/02/22/aspnet-core-allowlist-proxying-http-requests',
+  },
+  {
+    route: '/2019/02/22/whitelist-proxying-with-aspnet-core',
+    redirect: '/2019/02/22/aspnet-core-allowlist-proxying-http-requests',
+  },
+  {
+    route: '/2020/10/azure-devops-node-api-missing-episodes.html',
+    redirect: '/2020/10/31/azure-devops-node-api-git-api-getrefs-wiki-api',
+  },
+  {
+    route:
+      '/blog/2021-01-30-aspnet-serilog-and-application-insights/application-insights-properties.png',
+    redirect: '/assets/images/title-image-ba106de0d9e9a8f0299795d40b17e087.png',
+  },
+  {
+    route:
+      '/assets/images/permissioning-azure-pipelines-with-bicep-and-role-assignments-a540df332b847efc72e71362340d3b89.png',
+    redirect: '/assets/images/title-image-46cc8f39c57972b35cd9c539259b88f4.png',
+  },
+  {
+    route: '/2022/12/26/docusaurus-image-cloudinary-remark-plugin',
+    redirect: '/2022/12/26/docusaurus-image-cloudinary-rehype-plugin',
+  },
+  {
+    route: '/2021/03/15/from-blogger-to-docusaurus',
+    redirect: '/definitive-guide-to-migrating-from-blogger-to-docusaurus',
+  },
+];
+
+module.exports = brokenBacklinkGeneratedRedirects;
